% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beran_estimator.R
\name{beran.est}
\alias{beran.est}
\title{The ocation estimator of Beran (1974)}
\usage{
beran.est(x, theta, M, init, alpha)
}
\arguments{
\item{x}{An array of length n; represents the data.}

\item{theta}{A small number, should be of order \eqn{O_p(n^{-1/2})}. The default
is \eqn{4n^{-1/2}}. This is the tuning parameter for Fourier coefficient
estimation.}

\item{M}{The number of Fourier basis to be used to approximate \eqn{\phi_F}.}

\item{init}{Optional. A vector of initial estimators of m. The default value is the sample median.}

\item{alpha}{The confidence level for the confidence bands. An (1-alpha) percent
confidence interval is constructed. Alpha should lie in the interval (0, 0.50). The default
value is 0.05.}
}
\value{
A list of length two.   
\itemize{
\item \code{estimate:} An array of same length as init , giving the estimators of m
           based on the corresponding initial estimators in init. If init is missing,
             only one estimate is produced, which uses the sample median as the initial
             estimator.
\item \code{CI:} A matrix giving the (1-alpha) percent confidence intervals (CI). Each row corresponds to an initial estimator
                in init (if missing, the sample median is used). The first column
                corresponds to the lower CI, and the second column corresponds to the 
                upper CI.
}
}
\description{
Suppose the observations are sampled from a density \eqn{f(x-m)} where
m is the location parameter and f is an unknown density. This function, based on
Beran (1974), gives a nonparametric estimator of m. This estimator uses Fourier basis
expansion to estimate the  score function corresponding to the location model, which has the form
\deqn{\phi_F(t)=f'(F^{-1}(t))/f(F^{-1}(t)).}
  The Fourier
coefficients of the score function  are computed 
by \code{\link{score.coeff}}.
}
\details{
\code{theta:}   theta do not need to depend on the range of the data because
                        the estimators depend only on the rank of the dataponts.  If \eqn{theta=z_n} which 
                        equals \eqn{z n^{-1/2}}, then the estimated coefficients are 
                        root-n consistent by Theorem 2.1 of Beran (1974).

\code{M:}   A higher value of M decreases bias, but increases the variance. 
                     Theorem 4.1 of Beran (1974) states that if \eqn{M\to\infty} as
                     the sample size n grows   
                     and \eqn{\lim_{n\to \infty} M^6/n=0}, then this estimator is asymptotically efficient.
                     A larger M always gives a more conservative confidence interval.

\code{init:}  Beran (1974) recommends using sample median and warns that
                       the method will be sensitive to the choice of the preliminary estimator.
                       This should be a root-n consistent estimator of m.
                         Some other choices are the mean, or the trimmed mean.
}
\examples{
x <- rlogis(100); beran.est(x, M=10)
beran.est(x, M=20)
}
\references{
Beran, R. (1974). \emph{Asymptotically efficient adaptive 
rank estimates in location models}. Ann. Statist., 2, 63-74.
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu}.
}
