% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beran_selector.R
\name{beran.select}
\alias{beran.select}
\title{Tuning parameter selector for Beran (1974)'s location estimator}
\usage{
beran.select(x, t, Mvec, inth, B)
}
\arguments{
\item{x}{An array of length n; the dataset.}

\item{t}{Optional, an array of real numbers, contains values of theta, parameter needed for 
estimating the Fourier coefficient.}

\item{Mvec}{Optional, an array of integers, the number of basis functions to use. See details.}

\item{inth}{A number; the initial estimator. The default is the median.}

\item{B}{Optional, the number of bootstrap samples. The default is 100.}
}
\value{
A vector of foure numbers.
\itemize{
\item \code{param:} A 2 columns, the number of rows is same as the
                   length of inth. Each row gives the optimal
                   (dn, tn) for the corresponding inth value.
\item \code{estimte:}  Gives the estimators of \eqn{\theta}
      using the optimal (dn, tn)'s.
}
}
\description{
\code{\link{beran.est}} computes the location estimator in a location
shift model using Beran (1974)'s estimator. This function depends on
tuning parameters theta and M.  We estimate the MSE of the estimators and
to choose the best (theta, M) from a set of options.
}
\details{
To this end, we generate B Bootstrap samples and for each pair
of (theta, M), we estimate the MSE by computing
\deqn{\frac{\sum_{i=1}^B(\hat\theta_i(theta, M)-\theta)^2}{B}}
where \eqn{\theta_i(theta, M)} is the estimator based on the i-th bootstrap sample
and \eqn{\theta} is the true theta.

We take theta to be in c(0.01, seq(0.10, 0.80, by=0.05)) and take M to
be in (2, 4, 6, ..., 10).
}
\examples{
{}
  @export
}
\references{
Laha, N. \emph{ Location estimation fr symmetric and 
          log-concave densities}. Submitted.

Stone, C. (1975). \emph{Adaptive maximum likelihood estimators
of a location parameter}, Ann. Statist., 3, 267-284.
}
\seealso{
\code{\link{beran.est}}
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu}.
}
