% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stone_estimator.R
\name{giveth}
\alias{giveth}
\title{Stone (1975)'s estimator}
\usage{
giveth(x, inth, tn, dn)
}
\arguments{
\item{x}{A vector of length n; the dataset}

\item{inth}{An array giving some preliminary estimators for \eqn{\theta}. 
The function calculates different one-step estimators
for each of the preliminary estimators. The default value is
an array containing the mean, the median, and a trimmed mean (.05 from both tails).}

\item{tn}{A parameter associated with the bandwidth of the kernel
density estimator. The default value is 0.60.}

\item{dn}{A parameter for the truncation. The default value is 20.}
}
\value{
An array of the same length as inth. Each element is the one-step
          estimator based on the corresponding element in inth.
}
\description{
Gives a truncated one step estimator for  the location parameter in a symmetric
location family. This estimator is constructed using Stone (1975)'s
methods. This estimator uses the kernel density estimator (KDE) to estimate
the scores. Similar to all other one-step estimators, this estimator 
requires a preliminary estimator.
}
\details{
Stone(1975) takes inth to be the median in his simulations. The default values of dn and tn are taken from Stone (1975), who 
considered a sample of size 40. For asymptotic efficiency of the estimators,
a) dn\eqn{\to\infty} and tn\eqn{\to 0} b)
\deqn{\frac{dn^2}{n^{1-\epsilon}tn^6}=O(1)}
for some \eqn{\epsilon>0}. See (1.13) of Stone (1975) for more details.
}
\examples{
x <- rnorm(100); inth <- mean(x);
 giveth(x, inth=inth)
 giveth(x)
}
\references{
Laha, N. \emph{ Location estimation fr symmetric and 
          log-concave densities}. Submitted.

Stone, C. (1975). \emph{Adaptive maximum likelihood estimators
of a location parameter}, Ann. Statist., 3, 267-284.
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu}.
}
