% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{MLE_location}
\alias{MLE_location}
\title{The  MLE estimator of Laha (2020).}
\usage{
MLE_location(x, alpha = 0.05)
}
\arguments{
\item{x}{An array of length n; represents the data.}

\item{alpha}{The confidence level for the confidence bands. An (1-alpha) percent
confidence interval is constructed. Alpha should lie in the interval (0, 0.50). The default
value is 0.05.}
}
\value{
A list of length two.   
\itemize{
\item \code{estimate:}  A scalar.
\item \code{CI:} A vector of length two.
}
}
\description{
Suppose n univariate observations are sampled from a density \eqn{f(x-m)} where
m is the location parameter and f is an unknown symmetric density. This function computes
a one step estimator to estimte m. This estimator uses
the log-concave MLE estimator from the package \code{\link{logcondens.mode}} to estimate f.
}
\examples{
x <- rnorm(100); MLE_location(x)
}
\references{
Laha N. (2020). \emph{Location estimation for symmetric and
log-concave densities}. submitted.
}
\seealso{
\code{\link{p.mle}}, \code{\link{s.sym}}
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu}.
}
