% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxilliary_functions.R
\name{mle_score}
\alias{mle_score}
\title{Scores of log-concave MLE}
\usage{
mle_score(x)
}
\arguments{
\item{x}{a vector of length n; the dataset}
}
\value{
scores evaluated in between the data points, a vector of size n-1.
}
\description{
Calculates \deqn{\frac{\hat\phi_n(X_{i+1})-\hat\phi_n(X_i)}{X_{i+1}-X_i}}
for i=1,...,n., where \eqn{\hat\phi_n=\log\hat f_n}, the latter being the
unsmoothed log-concave MLE given by \code{\link{logcondens}}.
}
