% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partial_MLE_estimator.R
\name{p.mle}
\alias{p.mle}
\title{The partial MLE estimator of Laha (2020).}
\usage{
p.mle(x, init, q = 0, alpha = 0.05)
}
\arguments{
\item{x}{An array of length n; represents the data.}

\item{init}{Optional. An initial estimator of m. The default value is the sample median.}

\item{q}{A fraction between 0 and 1/2. Corresponds to the truncation parameter.
The default is 0, which indicates no truncation.}

\item{alpha}{The confidence level for the confidence bands. An (1-alpha) percent
confidence interval is constructed. Alpha should lie in the interval (0, 0.50). The default
value is 0.05.}
}
\value{
A list of length two.   
\itemize{
\item \code{estimate:}  A matrix of two columns, and the rownumber equals the length 
                       of q. The firs column is q, and the second column is estimates corresponding to q.
\item \code{CI:} A matrix of three columns, the first column is q,
the second column is the left point and the third column is the right point
of the confidence intervals corresponding to q.
}
}
\description{
Suppose n univariate observations are sampled from a density \eqn{f(x-m)} where
m is the location parameter and f is an unknown symmetric density. This function computes
a one step estimator to estimte m. This estimator uses
the log-concave MLE estimator from the package \code{\link{logcondens.mode}} to estimate f, and  is root-n consistent
for m provided f is log-concave.
}
\details{
\code{q:} If q is positive, the function
                     (1-2*q) percent observations from both tails while
                     computing the one step estimator. The scores are
                     estimated using the full data. See Laha et al. for more details.

\code{init:}  The default is mean. If init is the median, some jitter
                       (0.0001) is added.
}
\examples{
x <- rlogis(100); p.mle(x, q=c(0, 0.001, 0.01))
}
\references{
Laha N. (2020). \emph{Location estimation for symmetric and
log-concave densities}. submitted.
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu}.
}
