% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stone.selector.tuning.R
\name{stone.select}
\alias{stone.select}
\title{Tuning parameter selector for Stone (1975)'s location estimator}
\usage{
stone.select(x, inth, D, t, B)
}
\arguments{
\item{x}{An array of length n; the dataset.}

\item{inth}{A number; the initial estimator. The default is the median.}

\item{D}{Optional, an array of real numbers, contains values of D, parameter needed for tuning dn. See details.}

\item{t}{Optional, an array of real numbers, contains values of t, parameter needed for tuning tn. See details.}

\item{B}{Optional, the number of bootstrap samples. The default is 100.}
}
\value{
A list is returned:
\itemize{
\item \code{param:} A 2 columns, the number of rows is same as the
                   length of inth. Each row gives the optimal
                   (dn, tn) for the corresponding inth value.
\item \code{estimte:}  Gives the estimators of \eqn{\theta}
      using the optimal (dn, tn)'s.
\item\code{default_estimate:}A
}
}
\description{
\code{\link{giveth}} computes the location estimator in a location
shift model using Stone (1975)'s estimator. This function depends on
tuning parameters dn and tn.  We estimate the MSE of the estimators and
to choose the best (dn,tn) from a set of options.
}
\details{
To this end, we generate B Bootstrap samples and for each pair
of (dn, tn), we estimate the MSE by computing
\deqn{\frac{\sum_{i=1}^B(\hat\theta_i(dn,tn)-inth)^2}{B}}
where \eqn{\theta_i(dn,tn)} is the estimator based on the i-th bootstrap sample.

For asymptotic efficiency of the estimators,
a) dn\eqn{\to\infty} and tn\eqn{\to 0} b)
\deqn{\frac{(dn)^2}{n^{1-\epsilon}(tn)^6}=O(1)}
for some \eqn{\epsilon>0}. We take dn to be the minimum of
 \eqn{\min(D n^{1/2}(tn)^3 } and max(|x|)+2 (sd(x)*tn). Since dn is
 the truncation parameter and Stone (1975) uses kernel smoothening,
  we use the bound max(|x|)+2 (sd(x)tn).
and we take tn to be the minimum between one and \eqn{t n^{-1/7}/100}. 
We generate a set of (dn, tn) by varying
D and t acrros a grid. The default choice of the grid for t is c(0.01, seq(0.10, 0.80, by=0.05))
and that for D is
(0.5, 1, 1.5, 2, 3, 4,....., 20).
}
\examples{
{}
  @export
}
\references{
Laha, N. \emph{ Location estimation fr symmetric and 
          log-concave densities}. Submitted.

Stone, C. (1975). \emph{Adaptive maximum likelihood estimators
of a location parameter}, Ann. Statist., 3, 267-284.
}
\seealso{
\code{\link{giveth}}
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu}.
}
