% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stone.selector.tuning.R
\name{stone.select.ww}
\alias{stone.select.ww}
\title{Tuning parameter selector for Stone (1975)'s location estimator}
\usage{
stone.select.ww(x, inth)
}
\arguments{
\item{x}{An array of length n; the dataset.}

\item{inth}{A vector of initial estimators}

\item{D}{An array of real numbers, contains values of D, parameter needed for tuning dn. See details.}

\item{t}{An array of real numbers, contains values of t, parameter needed for tuning tn. See details.}
}
\value{
A list is returned:
\itemize{
\item \code{param:} A matrix with 2 columns, the number of rows is same as the
                   length of inth. Each row gives the optimal
                   (dn, tn) for the corresponding inth value.
\item \code{estimte:} An array of same length as inth. Gives the estimators of \eqn{\theta}
      using the optimal (dn, tn)'s.
}
}
\description{
\code{\link{giveth}} computes the location estimator in a location
shift model using Stone (1975)'s estimator. This function depends on
tuning parameters dn and tn.  We estimate the MSE of the estimators and
to choose the best (dn,tn) from a set of options.
}
\details{
To this end, we split the dataset in tan parts. Then we compute  estimators
\eqn{\hat \theta_i(dn,tn)} for each (dn,tn) under consideration from the i-th part
of the data, where i=1,...,10. We then estimate the MSE corresponding to each (dn, tn)
by computing
\deqn{\frac{\sum_{i=1}^10(\hat\theta_i(dn,tn)-\hat\theta(dn, tn))^2}{10}.}
We choose the (dn, tn) pairs which minimize the estimated MSE.

For asymptotic efficiency of the estimators,
a) dn\eqn{\to\infty} and tn\eqn{\to 0} b)
\deqn{\frac{(dn)^2}{n^{1-\epsilon}(tn)^6}=O(1)}
for some \eqn{\epsilon>0}. We take \eqn{dn=Dn^{1/2}(tn)^3}
and \eqn{tn=t n^{-1/7}}. We generate a set of (dn, tn) by varying
D and t in the set 
(0.5, 1, 1.5, 2, 3, 4,....., 20).
}
\examples{
x <- rnorm(100); inth <- mean(x);
 giveth(x, inth=inth)
 giveth(x)
}
\references{
Laha, N. \emph{ Location estimation fr symmetric and 
          log-concave densities}. Submitted.

Stone, C. (1975). \emph{Adaptive maximum likelihood estimators
of a location parameter}, Ann. Statist., 3, 267-284.
}
\seealso{
\code{\link{giveth}}
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu}.
}
