% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stone_estimator.R
\name{giveth}
\alias{giveth}
\title{Stone (1975)'s estimator}
\usage{
giveth(x, inthv)
}
\arguments{
\item{x}{A vector of length n; the dataset}

\item{dn}{A parameter for the truncation. The default value is 20.}

\item{tn}{A parameter associated with the bandwidth of the kernel
density estimator. The default value is 0.60.}

\item{inth}{A preliminary estimator for \eqn{\theta}. 
The default is the median.}

\item{alpha}{The confidence level for the confidence bands. An (1-alpha) percent
confidence interval is constructed. Alpha should lie in the interval (0, 0.50). The default
value is 0.05.}
}
\value{
A list of length two.   
\itemize{
\item \code{estimate:} An array of same length as init , giving the estimators of m
           based on the corresponding initial estimators in init. If init is missing,
             only one estimate is produced, which uses the sample median as the initial
             estimator.
\item \code{CI:} A matrix giving the (1-alpha) percent confidence intervals (CI). Each row corresponds to an initial estimator
                in init (if missing, the sample median is used). The first column
                corresponds to the lower CI, and the second column corresponds to the 
                upper CI.
}
}
\description{
Gives a truncated one step estimator for  the location parameter in a symmetric
location family. This estimator is constructed using Stone (1975)'s
methods. This estimator uses the kernel density estimator (KDE) to estimate
the scores. Similar to all other one-step estimators, this estimator 
requires a preliminary estimator.
}
\details{
Stone (1975) uses \eqn{r_n=\hat\sigma t_n} 
as the bandwidth of the Gaussian kernel, where \eqn{\hat\sigma} is the median of 
the \eqn{X_i-inth}'s.
For asymptotic efficiency of the estimators,
a) dn\eqn{\to\infty} and tn\eqn{\to 0} b)
\deqn{\frac{(dn)^2}{n^{1-\epsilon}(tn)^6}=O(1)}
for some \eqn{\epsilon>0}.  A rule of thumb plug-in estimate
for the optimal kernel width is 1.059\eqn{\hat\sigma n^{-1/5}} where \eqn{\hat\sigma}
is an estimator of the standard deviation, which does not satisfy the 
above relation. Stone(1975) takes inth to be the median in his simulations. The default choices of dn and tn 
are taken from Stone (1975), who 
considered a sample of size 40. Stone(1975)'s estimator uses Gaussian 
Kernels to estimate the unknown symmetric density.
}
\examples{
x <- rnorm(100); inth <- mean(x);
 giveth(x, inth=inth)
 giveth(x)
}
\references{
Laha, N. \emph{ Location estimation fr symmetric and 
          log-concave densities}. Submitted.

Stone, C. (1975). \emph{Adaptive maximum likelihood estimators
of a location parameter}, Ann. Statist., 3, 267-284.
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu}.
}
